ox.el: Allow the EXPORT_FILE_NAME to be generated
* ox.el (org-export-output-file-name): If EXPORT_FILE_NAME is enclosed
in backticks then evaluate the expression and use the returned value as
filename.

This patch encourages to reuse configuration (e.g. the document version
or title) in the export filename. A common use case when writing
documentation is that different versions of the document will be
exported over time. These are then distributed, e.g. via e-mail or
archived. By allowing the user to derive descriptive names during export
exchange of the document is greatly simplified.

E.g. the following excerpt will create a filename that contains the date
of export, the document title and the document version:

 #+TITLE: My wonderful document
 #+X-DOCUMENT-VERSION: 0.2-snapshot
 #+EXPORT_FILE_NAME: `(format "%s-%s (v%s)" (org-format-time-string "%Y-%m-%d_%Hh%Mm" (org-current-time)) (org-macro--find-keyword-value  "TITLE") (org-macro--find-keyword-value  "X-DOCUMENT-VERSION"))`